Added glob exclusion to ignore Vim backups directory
The ZSH system automatically loads all files with .zsh as configs and completion.sh as autocomplete configs. The */backups/* exclusion is an immediate fix for my Vim backups file location, as well as setting a pattern for other modules in the .dotfiles to save backups with their original name without worry of inclusion in the glob.